files relocation
