+++ lib/ui_screens/main_3_wished_and_seen/widget_tab_visited.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:places/domain_entities/place.dart';
 import 'package:places/ui_commons/enums.dart';
-import 'package:places/ui_commons/my_circle_progress_indicator.dart';
+import 'package:places/ui_widgets_commons/my_circle_progress_indicator.dart';
 import 'package:places/ui_commons/my_scroll_physics.dart';
 import 'package:places/ui_commons/ui_image_paths.dart';
 import 'package:places/ui_commons/ui_strings.dart';
